+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/Layer.java
@@ -251,7 +251,12 @@ public enum Layer {
      * The self observability of SkyWalking Go Agent,
      * which provides the abilities to measure the tracing performance and error statistics of plugins.
      */
-    SO11Y_GO_AGENT(41, true);
+    SO11Y_GO_AGENT(41, true),
+
+    /**
+     * Apache Flink is a framework and distributed processing engine for stateful computations over unbounded and bounded data streams
+     */
+    FLINK(42, true);
 
     private final int value;
     /**